sql/catalog/nstree: break up testing
Release note: None
