+++ pkg/sql/catalog/nstree/testdata/map/iterate
